+++ lib/model/settings/enums/accessibility_timeout.dart
@@ -8,18 +8,16 @@ extension ExtraAccessibilityTimeout on AccessibilityTimeout {
     switch (this) {
       case AccessibilityTimeout.system:
         return context.l10n.settingsSystemDefault;
-      case AccessibilityTimeout.appDefault:
-        return context.l10n.settingsDefault;
+      case AccessibilityTimeout.s1:
+        return context.l10n.timeSeconds(1);
       case AccessibilityTimeout.s3:
         return context.l10n.timeSeconds(3);
+      case AccessibilityTimeout.s5:
+        return context.l10n.timeSeconds(5);
       case AccessibilityTimeout.s10:
         return context.l10n.timeSeconds(10);
       case AccessibilityTimeout.s30:
         return context.l10n.timeSeconds(30);
-      case AccessibilityTimeout.s60:
-        return context.l10n.timeMinutes(1);
-      case AccessibilityTimeout.s120:
-        return context.l10n.timeMinutes(2);
     }
   }
 }