+++ lib/widgets/common/action_mixins/feedback.dart
@@ -106,24 +106,23 @@ mixin FeedbackMixin {
   }
 
   Future<Duration> _getSnackBarDuration(bool hasAction) async {
-    final appDefaultDuration = hasAction ? Durations.opToastActionDisplay : Durations.opToastTextDisplay;
     switch (settings.timeToTakeAction) {
       case AccessibilityTimeout.system:
-        final original = appDefaultDuration.inMilliseconds;
-        final millis = await (hasAction ? AccessibilityService.getRecommendedTimeToTakeAction(original) : AccessibilityService.getRecommendedTimeToRead(original));
-        return Duration(milliseconds: millis);
-      case AccessibilityTimeout.appDefault:
-        return appDefaultDuration;
+        if (hasAction) {
+          return Duration(milliseconds: await (AccessibilityService.getRecommendedTimeToTakeAction(Durations.opToastActionDisplay)));
+        } else {
+          return Duration(milliseconds: await (AccessibilityService.getRecommendedTimeToRead(Durations.opToastTextDisplay)));
+        }
+      case AccessibilityTimeout.s1:
+        return const Duration(seconds: 1);
       case AccessibilityTimeout.s3:
         return const Duration(seconds: 3);
+      case AccessibilityTimeout.s5:
+        return const Duration(seconds: 5);
       case AccessibilityTimeout.s10:
         return const Duration(seconds: 10);
       case AccessibilityTimeout.s30:
         return const Duration(seconds: 30);
-      case AccessibilityTimeout.s60:
-        return const Duration(minutes: 1);
-      case AccessibilityTimeout.s120:
-        return const Duration(minutes: 2);
     }
   }