+++ CHANGELOG.md
@@ -18,6 +18,7 @@ All notable changes to this project will be documented in this file.
 ### Changed
 
 - status and navigation bar transparency
+- default snack bar timeout to 3s
 - upgraded Flutter to beta v3.3.0-0.4.pre
 
 ### Fixed